refactor: deduplicate ${env.*} scanning and fix validation order
- Move extract_interpolation_refs to env_resolver.rs alongside
  interpolate so the ${...} scanning logic lives in one place.
  Both functions now share the same file with a doc comment noting
  they must stay in sync.
- Move self-reference check before validate_env_ref — structural
  errors should be caught before general ref validation.
- Move extract_interpolation_refs tests to env_resolver where the
  function now lives.
